+++ packages/browser/src/utils/event-utils.ts
@@ -280,7 +280,8 @@ export function getBrowserDetectionHints(): BrowserDetectionHints {
 
 export function getEventProperties(
     maskPersonalDataProperties?: boolean,
-    customPersonalDataProperties?: string[]
+    customPersonalDataProperties?: string[],
+    detectGoogleSearchApp?: boolean
 ): Properties {
     if (!userAgent) {
         return {}
@@ -290,12 +291,13 @@ export function getEventProperties(
         : []
     const [os_name, os_version] = detectOS(userAgent)
     const browserHints = getBrowserDetectionHints()
+    const browserOptions = { detectGoogleSearchApp }
 
     const properties = extend(
         stripEmptyProperties({
             $os: os_name,
             $os_version: os_version,
-            $browser: detectBrowser(userAgent, navigator.vendor, browserHints),
+            $browser: detectBrowser(userAgent, navigator.vendor, browserHints, browserOptions),
             $device: detectDevice(userAgent),
             $device_type: detectDeviceType(userAgent, {
                 // eslint-disable-next-line compat/compat
@@ -313,7 +315,7 @@ export function getEventProperties(
             $host: location?.host,
             $pathname: location?.pathname,
             $raw_user_agent: userAgent.length > 1000 ? userAgent.substring(0, 997) + '...' : userAgent,
-            $browser_version: detectBrowserVersion(userAgent, navigator.vendor, browserHints),
+            $browser_version: detectBrowserVersion(userAgent, navigator.vendor, browserHints, browserOptions),
             $browser_language: getBrowserLanguage(),
             $browser_language_prefix: getBrowserLanguagePrefix(),
             $screen_height: window?.screen.height,