+++ packages/convex/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @posthog/convex
 
+## 2.0.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/core@1.30.14
+  - posthog-node@5.36.8
+
 ## 2.0.18
 
 ### Patch Changes