+++ packages/webpack-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @posthog/webpack-plugin
 
+## 1.5.13
+
+### Patch Changes
+
+- Updated dependencies []:
+    - @posthog/core@1.30.14
+
+## 1.5.12
+
+### Patch Changes
+
+- [#3775](https://github.com/PostHog/posthog-js/pull/3775) [`018001c`](https://github.com/PostHog/posthog-js/commit/018001cb2e1417aae9a38bdc9338afce2e0d2c25) Thanks [@hpouillot](https://github.com/hpouillot)! - upgrade @posthog/cli version
+  (2026-06-09)
+
 ## 1.5.11
 
 ### Patch Changes